# personal-finance-management/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RESTful API Personal Finance Management",
    "description": "OpenAPI for Personal Finance Management, with the following features:\n1. User Management\n2. User Income\n3. User Expenses\n3. User Savings\n4. Personal Finance User Dashboard",
    "version": "1.0",
    "contact": {
      "name": "Mahfudz Nur Fadhila",
      "email": "mahfudnf63@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api"
    }
  ],
  "paths": {
...
    "/expenses" : {
      "post": {
        "tags": ["expense"],
        "summary": "Create expense",
        "description": "Create new expense",
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json" : {
              "schema": {
                "type": "object",
                "required": ["amount","category"],
                "properties": {
                  "amount" : {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "example": 50000
                  },
                  "category" : {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "transport"
                  },
                  "description" : {
                    "type": "string",
                    "maxLength": 500,
                    "example": "pengeluaran transport"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200" : {
            "description": "Create expense success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateIncomeSuccess"
                },
                "example": {
                  "data" : {
                    "id": "11334555",
                    "amount": 50000,
                    "category": "transport",
                    "createdAt": "2026-02-25",
                    "description": "pengeluaran transport"
                  }
                }
              }
            }
          },
          "401" : {
            "description": "Create expense failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["expense"],
        "summary": "List expenses",
        "description": "Retrieve data according to the query parameters entered by the user",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "responses": {
          "200" : {
            "description": "Success get List expenses",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseListIncome"
                },
                "example": {
                  "data" :[
                    {
                      "id": "11334555",
                      "amount": 50000,
                      "category": "transport",
                      "createdAt": "2026-02-25",
                      "description": "pengeluaran transport"
                    }
                  ],
                "paging" : {
                  "currentPage" : 0,
                  "totalPages" : 10,
                  "size" : 10
                }
                }
              }
            }
          },
          "401" : {
            "description": "Get list expenses failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/{expenseId}" : {
      "get": {
        "tags": ["expense"],
        "summary": "Get expense",
        "description": "Retrieve expense based on id",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "responses": {
          "200" : {
            "description": "Get expense success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateIncomeSuccess"
                },
                "example": {
                  "data" : {
                    "id": "11334555",
                    "amount": 50000,
                    "category": "transport",
                    "createdAt": "2026-02-25",
                    "description": "pengeluaran transport"
                  }
                }
              }
            }
          },
          "404" : {
            "description": "Get expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                },
                "example": {
                  "errors" : "Expense not found"
                }
              }
            }
          },
          "401" : {
            "description": "Get expense failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["expense"],
        "summary": "Edit expense",
        "description": "Edit expenses based on income id in path variable",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json" : {
              "schema": {
                "properties": {
                  "amount" : {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 0,
                    "example": 70000
                  },
                  "category" : {
                    "type": "string",
                    "maxLength": 100,
                    "example": "listrik"
                  },
                  "description" : {
                    "type": "string",
                    "maxLength": 500,
                    "example": "pengeluaran bayar listrik"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200" : {
            "description": "Edit expense success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseCreateIncomeSuccess"
                },
                "example": {
                  "data" : {
                    "id": "11334555",
                    "amount": 70000,
                    "category": "listrik",
                    "createdAt": "2026-02-25",
                    "description": "pengeluaran bayar listrik"
                  }
                }
              }
            }
          },
          "404" : {
            "description": "Get expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                },
                "example": {
                  "errors" : "Expense not found"
                }
              }
            }
          },
          "401" : {
            "description": "Edit expense failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["expense"],
        "summary": "Delete expense",
        "description": "Delete expenses based on income id in path variable",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth" : []
          }
        ],
        "responses": {
          "200" : {
            "description": "Delete expense success",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/responseDataOk"
                }
              }
            }
          },
          "404" : {
            "description": "Delete expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                },
                "example": {
                  "errors" : "Expense not found"
                }
              }
            }
          },
          "401" : {
            "description": "Delete expense failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseUnauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth" : {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "responseDataOk" : {
        "type": "object",
        "properties": {
          "data" : {
            "type": "string",
            "example": "OK"
          }
        }
      },
...
      "responseUnauthorized" : {
        "type": "object",
        "properties": {
          "errors": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
      "responseCreateIncomeSuccess" : {
        "type": "object",
        "properties": {
          "data" : {
            "type": "object",
            "properties": {
              "incomeId" : {
                "type": "string",
                "example": "1231313"
              },
              "amount" : {
                "type": "integer",
                "format": "int64",
                "example": 100000
              },
              "category" : {
                "type": "string",
                "example": "gaji"
              },
              "createdAt" : {
                "type": "string",
                "format": "date",
                "example": "2026-02-23"
              },
              "description" : {
                "type": "string",
                "example": "pendapatan gaji bulan juni"
              }
            }
          }
        }
      },
      "paging" : {
        "type": "object",
        "properties": {
          "currentPage" : {
            "type": "integer",
            "example": 0
          },
          "totalPages" : {
            "type": "integer",
            "example": 10
          },
          "size" : {
            "type": "integer",
            "example": 10
          }
        }
      },
      "responseListIncome" : {
        "type": "object",
        "properties": {
          "data" : {
            "type": "array",
            "items": {
              "properties": {
                "incomeId" : {
                  "type": "string",
                  "example": "1231313"
                },
                "amount" : {
                  "type": "integer",
                  "format": "int64",
                  "example": 100000
                },
                "category" : {
                  "type": "string",
                  "example": "gaji"
                },
                "createdAt" : {
                  "type": "string",
                  "format": "date",
                  "example": "2026-02-23"
                },
                "description" : {
                  "type": "string",
                  "example": "pendapatan gaji bulan juni"
                }
              }
            }
          },
          "paging" : {
            "$ref": "#/components/schemas/paging"
          }
        }
      },
...
    }
  }
}